The Python implementation of Silhouette coefficient
The Silhouette coefficient is one of the most widely used evaluation metrics for clustering algorithms. Silhouette is a method of interpreting and validating consistency within data clusters. It is a measure of how similar an object is to its own cluster (cohesion) in comparison to other clusters (separation).The score of Silhouette Coefficient is bounded between -1 for incorrect clustering and +1 for highly dense clustering. Scores around zero indicate overlapping clusters.